src/services/capstone 2/requirements.txt
flask

google-auth

google-auth-oauthlib

google-api-python-client

werkzeug

requests
